Update MailFetcher.php
Added condition if MAIL_DELETE is set to true in HELPDESK it will delete mails after fetching, if is set to anything else it will do nothing

+++ hdz/app/Libraries/MailFetcher.php
@@ -71,7 +71,9 @@ class MailFetcher
                             }
                         }
                     }
-                    $mailbox->deleteMail($mail->id);
+                  if(Helpdesk::MAIL_DELETE == true){
+                      $mailbox->deleteMail($mail->id);
+                  }
                 }
                 $mailbox->disconnect();
             }
